feature: Auto update weather data
Weather data is updated on application startup.
In addition, the application will hourly poll the WFS during the morning
until it receives the morning issuance. Then it will wait until the
afternoon before polling for the afternoon issuance.

Added a deduplication procedure to schema-postgres.sql which will delete
duplicate weather forecast features that result from polling the WFS.

TODO check that it works over the course of an actual day