feature: added gas endpoint to return gas stations along a route.
To make a GET call to the api/route/gas endpoint with the specified parameters, you would need to pass the following query parameters:

departureLat: The latitude of the departure location.
departureLng: The longitude of the departure location.
arrivalLat: The latitude of the arrival location.
arrivalLng: The longitude of the arrival location.
travelersMeterCapacity: The capacity of the traveler's meter (presumably for gas).
type: The type of gas station you are looking for.
TODO: Unit Tests