fix: Speed up check_route_weather by using view with subdivided geometry
Subdividing the geometry makes better use of the geometry index at the
cost of more storage space and having to deduplicate the results.

A bit inefficient as the view is refreshed by subdividing all
weather features instead of subdividing newly added features.
This might be solvable through Incremental View Maintenance, but this is
not available yet.

2.465 seconds for 1125 segments without a view to 0.204 seconds with a
view.